Simplify image resolver code
