+++ .changeset/red-onions-exist.md
@@ -1,0 +1,5 @@
+---
+'@keystone-next/fields': patch
+---
+
+Simplified image input resolver.
